pyproject.toml
[tool.poetry]
name = "accounting-paper-generator"
version = "0.1.0"
description = ""
authors = ["Laurence Hook <hookla1976@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.38.0"
langchain = "^0.3.0"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
langchain-anthropic = "^0.2.1"
black = "^24.8.0"
isort = "^5.13.2"
pylint = "^3.3.0"
flake8 = "^7.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
